requirements.txt
gradio ~= 4.41.0
gitpython